Implement RGB to Y'CbCr transforms.
First attempt to write a set of functions to transform to and from
Y'CbCr (and the constant luminance Yc'Cbc'Crc' special case for
Rec.2020).

Move ycbcr.py into rgb sub-folder. As Y'CbCr is a method of encoding RGB values, rather than a colour space as such, it is more suitably placed there.

Fix typo in example. Missing ' in YcCbcCrc_to_rgb example.