PR: Datasets: TCS: Adding TCS15 (#1337)
Datasets: TCS: Adding TCS15

TCS15 was added to the CIE 1995 test color samples to get a better
representation of the Japanese skin complexion.
This commit adds this color sample and adjust the CRI test file
to pass.

Note, that the TCS15 dataset is published from 380-780 nm in 5 nm increments.
This differs from all other TCS data, which is published from 360-830 nm
(also in 5 nm increments).

The integration of TCS15 in this commit is breaking existing code,
which can be observed in the following regressions:
* FAILED colour/colorimetry/tests/test_uniformity.py::TestSpectralUniformity::test_spectral_uniformity
* FAILED colour/colorimetry/uniformity.py::colour.colorimetry.uniformity.spectral_uniformity

The integration is addressed in a following commit.

Source of data is the CIE's open access dataset page:
https://www.cie.co.at/datatable/spectral-radiance-factors-test-colour-sample-15-japanese-skin-complexion-5nm-wavelength

Data source reference:
CIE 2024, Spectral radiance factors of test-colour sample #15 of the Japanese
skin complexion, 5nm wavelength steps, International Commission on Illumination
(CIE), Vienna, AT, DOI: 10.25039/CIE.DS.7chm7z5h

Signed-off-by: Christoph Müllner <christophm30@gmail.com>

* CRI: Fix integration of TCS15

This commit fixes the TCS15 integration with the following steps:
* Adding COLOUR_RENDERING_INDEX_METHODS to switch between "CIE 1995" and
  "CIE 2024" TCS sets
* Introduce `SDS_TCS_SETS` to get the TCS set for the desired method
* Introduce SDS_TCS_CIE_1995 (was SDS_TCS before) SDS_TCS_CIE_2024
* Add new parameter `method` to colour.colour_rendering_index to select
  the TCS set
* Add a test for this new `method` parameter
* Adjust the docs to reflect these changes with inspiration from
  the symbols exported by the CQS module

Suggested-by: Thomas Mansencal <thomas.mansencal@gmail.com>
Signed-off-by: Christoph Müllner <christophm30@gmail.com>

---------

Signed-off-by: Christoph Müllner <christophm30@gmail.com>